@@ -14,42 +14,42 @@
     "sentry:sourcemaps": "sentry-cli sourcemaps inject --org alexander-grattan --project artificial-unintelligence-server ./dist && sentry-cli sourcemaps upload --org alexander-grattan --project artificial-unintelligence-server ./dist"
   },
   "dependencies": {
-    "@sentry/node": "^8.20.0",
-    "@sentry/profiling-node": "^8.20.0",
-    "cookie-parser": "^1.4.6",
+    "@sentry/node": "^10.17.0",
+    "@sentry/profiling-node": "^10.17.0",
+    "cookie-parser": "^1.4.7",
     "cors": "^2.8.5",
     "database": "workspace:*",
-    "dotenv": "^16.4.5",
-    "drizzle-orm": "^0.32.1",
-    "express": "^4.19.2",
-    "helmet": "^7.1.0",
-    "ioredis": "^5.4.1",
-    "morgan": "^1.10.0",
-    "replicate": "^0.32.0",
-    "socket.io": "^4.7.5"
+    "dotenv": "^17.2.3",
+    "drizzle-orm": "^0.44.6",
+    "express": "^5.1.0",
+    "helmet": "^8.1.0",
+    "ioredis": "^5.8.0",
+    "morgan": "^1.10.1",
+    "replicate": "^1.2.0",
+    "socket.io": "^4.8.1"
   },
   "devDependencies": {
-    "@jest/globals": "^29.7.0",
-    "@sentry/cli": "^2.33.0",
-    "@types/cookie-parser": "^1.4.7",
-    "@types/cors": "^2.8.17",
-    "@types/express": "^4.17.21",
-    "@types/morgan": "^1.9.9",
-    "@types/node": "^20.14.12",
-    "cross-env": "^7.0.3",
+    "@jest/globals": "^30.2.0",
+    "@sentry/cli": "^2.56.0",
+    "@types/cookie-parser": "^1.4.9",
+    "@types/cors": "^2.8.19",
+    "@types/express": "^5.0.3",
+    "@types/morgan": "^1.9.10",
+    "@types/node": "^24.6.2",
+    "cross-env": "^10.1.0",
     "eslint-config-custom": "workspace:*",
-    "jest": "^29.7.0",
-    "nodemon": "^3.1.4",
-    "ts-jest": "^29.2.3",
+    "jest": "^30.2.0",
+    "nodemon": "^3.1.10",
+    "ts-jest": "^29.4.4",
     "ts-node": "^10.9.2",
     "tsconfig": "workspace:*",
-    "typescript": "5.4.4"
+    "typescript": "5.9.3"
   },
   "engines": {
-    "node": "20.16.0"
+    "node": "22.14.0"
   },
   "optionalDependencies": {
-    "bufferutil": "^4.0.8",
-    "utf-8-validate": "^6.0.4"
+    "bufferutil": "^4.0.9",
+    "utf-8-validate": "^6.0.5"
   }
 }
